API

External API

Internal API

Base extensions

External API

SMGReader.AuxField — Type

AuxField

Describes the location and type of an auxiliary field in a BAM record.

Fields:

tag::NTuple{2, UInt8}: Two-character tag identifying the field.

typechar::UInt8: Type character for the field (e.g., 'A', 'Z', 'B', etc.).

elemtypechar::UInt8: Element type character for array fields ('B').

start::Int: Start index (1-based) of the field value in the record's byte array.

stop::Int: Stop index (1-based) of the field value in the record's byte array.

SMGReader.AuxFieldIter — Type

AuxFieldIter(record::BamRecord)

An iterator over the auxiliary fields of a BAM record. Yields AuxField objects describing the location and type of each aux field in the record.

SMGReader.AuxMapMod — Type

AuxMapMod()

Construct an map of the aux data for HP, MM, ML fields. For standard use in chromatin stencilling (non-fire) and direct RNA.

SMGReader.AuxMapModFire — Type

AuxMapModFire()

Construct and map of the aux data for a FIRE BAM/CRAM mapping fields

hp: Haplotype

mm: Run length encoded mod bases

ml: Mod Probability

ns: Nucleosome positions

nl: nucleosome length

as: msp positions

al: msp lengths

aq: msp quality score

SMGReader.DirectRNA — Type

DirectRNA(auxmap::AuxMap)

Construct preallocated HTS read data building a read to genome map for DirectRNA setting an BAM auxillary data map AuxMap. e.g.

For generic tracking of modifications DirectRNA(AuxMapMod())

SMGReader.DirectRNAAlignBlocks — Type

DirectRNAAlignBlocks(auxmap::AuxMap)

Construct preallocated HTS read data building both a read to genome map and an align blocks to genome map for spliced alignments. e.g.

For generic tracking of modifications DirectRNAAlignBlocks(AuxMapMod())

SMGReader.ModIterator — Type

struct ModIterator

An iterator over all modification calls (as ModificationInfo) in a BAM/CRAM record, as described by the MM and ML auxiliary fields.

Fields

record::BamRecord: The BAM/CRAM record.

mmaux::AuxField: The MM auxiliary field (modification code and run length encoding of).

mlaux::AuxField: The ML auxiliary field (modification probabilities).

SMGReader.ModIterator — Method

ModIterator(record::BamRecord, rdata::HTSReadData)

Construct a ModIterator for a given BAM/CRAM record, using the mm and ml fields in the auxmap

SMGReader.ModIterator — Method

ModIterator(record::BamRecord)

Construct a ModIterator for a given BAM/CRAM record, automatically extracting the MM and ML fields.

SMGReader.Modification — Type

Modification

Enumerates supported DNA/RNA modification types, such as 6mA, 5mC, 5hmC, inosine, pseudouridine, and others. Used to identify the type of modification parsed from MM/ML fields.

Currently supported mod codes:

Enum value	Modification	BAM code

mod_6mA	N6-methyladenine	a

mod_5mC	5-methylcytosine	m

mod_5hmC	5-hydroxymethylcytosine	h

mod_inosine	Inosine	17596

mod_pseU	Pseudouridine	17802

mod_4mC	N4-methylcytosine	21839

SMGReader.ModificationInfo — Type

ModificationInfo

Holds information about a single modification call parsed from a BAM/CRAM record.

Fields

mod::Modification: The modification type.

strand::Bool: true for forward (+), false for reverse (-) strand.

pos::Int: 1-based position in the read sequence.

prob::UInt8: Modification probability (from ML field).

SMGReader.StencillingData — Type

StencillingData(auxmap::AuxMap)

Construct preallocated HTS read data building a read to genome map for chromatin stencilling setting an BAM auxillary data map AuxMap. e.g.

For generic tracking of modifications StencillingData(AuxMapMod())

For tracking FIRE fields StencillingData(AuxMapModFire())

SMGReader.VectorBuffer — Type

mutable struct VectorBuffer{T} <: AbstractVector{T}

A resizable buffer that wraps a Vector{T} and tracks a logical length (len). Efficient for repeated resizing and reuse, e.g., when processing records in a loop.

Recommendation is to allocate a large enough vector buffer to work with, e.g.
following constants are set in processedread.jl

const MOD_6MA_BUFFER_LENGTH = 70_000
const MOD_5MC_BUFFER_LENGTH = 70_000
const READ_BUFFER_LENGTH = 200_000
const ALIGN_BLOCK_LENGTH = 5_000

SMGReader.VectorBuffer — Method

VectorBuffer{T}(backcapacity::Int, initcapacity=0)

Construct a VectorBuffer with backing storage of size backcapacity and logical length initcapacity.

SMGReader.cigarstring — Method

cigarstring(record)

Return CIGAR string of record.

SMGReader.cigarvec — Function

cigarvec(record)

Get CIGAR description of alignment of record returned as lengths (Vector{Int}) and codes (Vector{Char}).

To build alignment map use readtogenome instead of working with cigar directly

SMGReader.eachintersection — Method

eachintersection(reader::HTSFileReader, chrom::AbstractString, loc::UnitRange{Int})

Construct an iterator of reader over chrom and loc.

SMGReader.eachrecord — Method

eachrecord(reader::HTSFileReader)

Construct an iterator of reader in record order. Multiple successive calls will continue iterating from current point in reader, resetting iterator requires closing and opening file.

SMGReader.firegenome — Method

firegenome(record::BamRecord, recorddata::StencillingData{AuxMapModFire}

Construct a function for mapping fire nucleosomes and MSP coordiates from read coordinates to genome filtering out entries that do not map

Usage: firegenomecoords = firegenome(record, recorddata)

for (pos, len, qual) in firegenomcoords(firemsps(record, recorddata))
    ### do something with genome coordinates
end

SMGReader.firegenomecoords — Method

firegenomecoords(pos::Int, len::Int, record::BamRecord, recorddata::HTSReadData; onebased=true)

Special version of genomecoords for fiber-tools and fire elements that are encoded 0-based.

pos is zero-based position on the read

one-based = true returns one based inclusive coords

one-based = false returns zero based exclusive coords

SMGReader.firemsps — Method

firemsps(r::BamRecord, recorddata::StencillingData{AuxMapModFire})

Construct iterator of fire msp positions, lengths, and quality scores in read coordinates Usage:

for (pos, len, qual) in firemsps(record, recorddata)
    ### do something 
end

SMGReader.firenucs — Method

firenucs(r::BamRecord, recorddata::StencillingData{AuxMapModFire})

Construct iterator of fire nucleosome positions, lengths in read coordinates Usage:

for (pos, len) in firenucs(record, recorddata)
    ### do something 
end

SMGReader.flag — Method

flag(record::BamRecord)

Get flag of record

SMGReader.genomecoords — Method

genomecoords(pos::Int, len::Int, record::BamRecord, recorddata::HTSReadData; onebased=true)

Map read coordinates to genome coordates that uses alignmap in recorddata,

pos is one-based to match the alignment map

one-based = true returns one based inclusive coords

one-based = false returns zero based exclusive coords

SMGReader.genomecoords — Method

genomecoords(pos::Int, record::BamRecord, recorddata::HTSReadData; onebased=true)

Map read coordinates to genome coordates that uses alignmap in recorddata, onebased=true for 1-based coordinates

SMGReader.haplotype — Function

haplotype(record::BamRecord, recorddata::HTSReadData)

Return haplotype of record and nothing if haplotype missing

SMGReader.hashaplotype — Method

hashaplotype(record::BamRecord, recorddata::HTSReadData)

Return true if record has haplotype field

SMGReader.indexfile — Method

indexfile(file)

Return a .bai or .crai file as appropriate for bam and cram files

SMGReader.ispositive — Method

ispositive(record::BamRecord)

Determine if record is aligned to the positive strand

SMGReader.leftposition — Method

leftposition(record::BamRecord)

Return the position of the record in 0-based coords

SMGReader.mappingquality — Method

mappingquality(record::BamRecord)

Return mapping quality of record

SMGReader.matepos — Method

mateposition(record::BamRecord)

return the mate position in 0-based coords

SMGReader.materefname — Method

materefname(reader::HTSFileReader, record::BamRecord)

Return the reference name of the mate of record read by reader. reader is necessary to load the ref name from the header of the bam/cram file.

SMGReader.processread! — Method

processread!(record::BamRecord, recorddata::HTSReadData)

Process read mapping auxillary fields and building read to genome map using preallocated recorddata.

SMGReader.qname — Method

qname(record::BamRecord)

Return query name of record. The read name.

SMGReader.qual — Method

qual(record::BamRecord)

Get quality scores as byte array

SMGReader.querylength — Method

querylength(record::BamRecord)

Return the query length of the record

SMGReader.refname — Method

refname(reader::HTSFileReader, record::BamRecord)

Return the reference name of record read by reader. reader is necessary to load the ref name from the header of the bam/cram file.

SMGReader.rightposition — Method

rightposition(recorddata::HTSReadData)

Get the right position of alignment on genome.

SMGReader.seq — Method

seq(record)

Return seq string of record.

SMGReader.setlength! — Function

setlength!(buff::VectorBuffer, newlen, backfactor=1.2)

Resize the logical length of the buffer to newlen. If the backing storage is too small, resize it by a factor of backfactor or double the current length, whichever is larger. Returns the buffer.

SMGReader.templatelength — Method

templatelength(record::BamRecord)

return the template length of the record

SMGReader.validflag — Method

validfrag(record::BamRecord)

Return true if flag is 0 or 16.

Internal API

SMGReader.ModIteratorState — Type

ModIteratorState

Mutable internal state of ModIterator

Fields

newrecord::Bool: Bool marking whether next iteration will return a new modification with different modcode

base::UInt8: Current base associated with modcode

strand::Bool: String of modfication

mod::Modification: Current modification

mm_pos::Int: Position in the MM string

ml_pos::Int: Position in the ML vector

seq_pos::Int: Position in the bam sequence

SMGReader.alignmentblocks! — Method

alignmentblocks!(record::BamRecord, alignblocks::VectorBuffer{AlignBlock})

In place update of alignment blocks for a record

SMGReader.auxmap! — Method

auxmap!(record::BamRecord, auxmap<:Auxmap)

Construct map of auxillary data from record identifying fields specified by auxmap.

SMGReader.checkbounds — Method

checkbounds(buff::VectorBuffer, i)

Check if index i is within the logical bounds of the buffer. Throws BoundsError if out of bounds.

SMGReader.findnextbase — Method

findnextbase(record::BamRecord, base::UInt8, start::Int, skip::Int)

Find the next occurence of base = (0x01, 0x02, 0x04, 0x08)for(A, C, G, T)within the bam bit packed sequence, starting from positionstartskipping overskip` occurences of the base.

This function is designed to locate the next base for which modification data is available for parsing the MM run length encoded aux field.

SMGReader.findprevbase — Method

findprevbase(record::BamRecord, base::UInt8, start::Int, skip::Int)

Find the prev occurence of base = (0x01, 0x02, 0x04, 0x08)for(A, C, G, T)within the bam bit packed sequence, starting from positionstartskipping overskip` occurences of the base.

This function is designed to locate the prev base for which modification data is available for parsing the MM run length encoded aux field.

This is the reverse version of findnextbase for use in reverse strand aligned reads.

SMGReader.firemsplen — Method

firemsplen(r::BamRecord, recorddata::StencillingData{AuxMapModFire})

Return fiber tools annotated MSP lengths on read coordinates.

SMGReader.firemsppos — Method

firemsppos(r::BamRecord, recorddata::StencillingData{AuxMapModFire})

Return fiber tools annotated MSP positions on read coordinates. 0-based coordinates.

SMGReader.firemspqual — Method

firemspqual(r::BamRecord, recorddata::StencillingData{AuxMapModFire})

Return fiber tools annotated MSP quality scores on read coordinates.

SMGReader.firenuclen — Method

firenuclen(r::BamRecord, recorddata::StencillingData{AuxMapModFire})

Return fiber tools annotated nucleosome lengths on read coordinates.

SMGReader.firenucpos — Method

firenucpos(r::BamRecord, recorddata::StencillingData{AuxMapModFire})

Return fiber tools annotated nucleosome positions on read coordinates 0-based coordinates.

SMGReader.getmmml — Method

getmmml(record::BamRecord)

Function to retrieve location of MM and ML auxillary fields as AuxFields

SMGReader.getmodbasestrand — Method

getmodbasestrand(b::UInt8, ispos::Bool) -> UInt8

Given a base character and strand, return the encoded base for modification search.

SMGReader.parse_mod_code — Method

parse_mod_code(data, i)

Parse the modcode at position i in bytearray data Return (mod, next_i) where mod is a Modifcation and next_i is the next index after the modcode in the bytearray.

SMGReader.readtogenome! — Method

readtogenome!(record::BamRecord, alignmap::VectorBuffer{T}) where {T}

In place update of alignmap VectorBuffer with basepair level read to genome map. Uses CIGAR information in record

Base Extensions

Base.close — Method

close(reader::HTSFileReader)

Close open reader.

Base.getindex — Method

Base.getindex(buff::VectorBuffer{T}, inds::AbstractVector{Bool})

Return a regular array of elements where inds is true. Logical indexing must match the logical length of the buffer.

Base.getindex — Method

Base.getindex(buff::VectorBuffer{T}, i::Int)

Return the element at logical index i in the buffer.

Base.getindex — Method

Base.getindex(buff::VectorBuffer{T}, r::UnitRange{Int})

Return a slice of the buffer for the given range r.

Base.iterate — Method

Base.iterate(iter::AuxFieldIter)
Base.iterate(iter::AuxFieldIter, i)

Iterate over the auxiliary fields in a BAM record. Returns a tuple (AuxField, next_index) for each field, or nothing when done.

Base.iterate — Method

Base.iterate(iter::ModIterator)

Iterate through modifications in order they are present in BAM record returning `ModInfo` for each.

Base.open — Method

open(HTSFileReader, path; idx_path=indexfile(path), bamthreads=Threads.nthreads())

Use ccall to hts_open the file at path. - idx_path gives location of index and is mandatory - bamthreads specifies number of threads used by htslib

Base.read! — Method

Base.read!(reader::HTSFileReader, record::BamRecord)

Read next record in file order from reader and in place store result in record.

Base.read — Method

Base.read(reader::HTSFileReader)

Allocating version of read! Read next record in file order from reader return a record, recommended to use read!.

Base.setindex! — Method

Base.setindex!(buff::VectorBuffer{T}, val, i::Int)

Set the element at logical index i to val.